Refactor graph engine tests
